+++ src/hooks/useResultActions.ts
@@ -1,0 +1,170 @@
+import { useEffect, useRef, useState } from 'react';
+import type { useRouter } from 'expo-router';
+
+import { showInterstitial } from '@/src/ads/interstitial';
+import { useHighScore } from '@/src/hooks/useHighScore';
+import type { GameState } from '@/src/game/state';
+import type { MazeData } from '@/src/types/maze';
+
+interface Options {
+  state: GameState;
+  maze: MazeData;
+  nextStage: () => void;
+  resetRun: () => void;
+  router: ReturnType<typeof useRouter>;
+  showSnackbar: (msg: string) => void;
+  pauseBgm: () => void;
+  resumeBgm: () => void;
+}
+
+/**
+ * リザルト表示やメニュー関連の状態を扱うフック
+ */
+export function useResultActions({
+  state,
+  maze,
+  nextStage,
+  resetRun,
+  router,
+  showSnackbar,
+  pauseBgm,
+  resumeBgm,
+}: Options) {
+  const {
+    highScore,
+    newRecord,
+    setNewRecord,
+    updateScore,
+  } = useHighScore(state.levelId);
+
+  const [showResult, setShowResult] = useState(false);
+  const [gameOver, setGameOver] = useState(false);
+  const [stageClear, setStageClear] = useState(false);
+  const [gameClear, setGameClear] = useState(false);
+  const [showMenu, setShowMenu] = useState(false);
+  const [debugAll, setDebugAll] = useState(false);
+
+  const [okLocked, setOkLocked] = useState(false);
+  const okLockedRef = useRef(false);
+
+  // ゴール到達や捕まったときの処理をまとめる
+  useEffect(() => {
+    const willChangeMap = state.stage % maze.size === 0;
+    if (state.pos.x === maze.goal[0] && state.pos.y === maze.goal[1]) {
+      setStageClear(true);
+      setGameOver(false);
+      setGameClear(state.finalStage);
+      setShowResult(true);
+      setDebugAll(willChangeMap);
+      if (state.levelId) {
+        const current = {
+          stage: state.stage,
+          steps: state.steps,
+          bumps: state.bumps,
+        };
+        updateScore(current, state.finalStage);
+      } else {
+        setNewRecord(false);
+      }
+    } else if (state.caught) {
+      setGameOver(true);
+      setStageClear(false);
+      setShowResult(true);
+      setDebugAll(true);
+      if (state.levelId) {
+        const current = {
+          stage: state.stage - 1,
+          steps: state.steps,
+          bumps: state.bumps,
+        };
+        updateScore(current, false);
+      } else {
+        setNewRecord(false);
+      }
+    }
+  }, [
+    state.pos,
+    state.caught,
+    maze.goal,
+    state.finalStage,
+    state.stage,
+    maze.size,
+    state.steps,
+    state.bumps,
+    state.levelId,
+    updateScore,
+    setNewRecord,
+  ]);
+
+  // OK ボタン押下時の処理
+  const handleOk = async () => {
+    if (okLockedRef.current) return;
+    okLockedRef.current = true;
+    setOkLocked(true);
+    if (gameOver) {
+      resetRun();
+    } else if (gameClear) {
+      resetRun();
+      router.replace('/');
+    } else if (stageClear) {
+      if (state.stage % 9 === 0 || state.stage === 1) {
+        try {
+          pauseBgm();
+          await showInterstitial();
+        } catch (e) {
+          console.error('interstitial error', e);
+          showSnackbar('広告を表示できませんでした');
+        } finally {
+          resumeBgm();
+        }
+      }
+      nextStage();
+    }
+    setShowResult(false);
+    setGameOver(false);
+    setDebugAll(false);
+    setStageClear(false);
+    setGameClear(false);
+    setNewRecord(false);
+    okLockedRef.current = false;
+    setOkLocked(false);
+  };
+
+  // リセット処理
+  const handleReset = () => {
+    setShowMenu(false);
+    setGameOver(false);
+    setStageClear(false);
+    setGameClear(false);
+    setNewRecord(false);
+    resetRun();
+  };
+
+  // タイトルへ戻る処理
+  const handleExit = () => {
+    setShowMenu(false);
+    setGameOver(false);
+    setStageClear(false);
+    setGameClear(false);
+    setNewRecord(false);
+    resetRun();
+    router.replace('/');
+  };
+
+  return {
+    showResult,
+    gameOver,
+    stageClear,
+    gameClear,
+    highScore,
+    newRecord,
+    showMenu,
+    setShowMenu,
+    debugAll,
+    setDebugAll,
+    okLocked,
+    handleOk,
+    handleReset,
+    handleExit,
+  } as const;
+}
